requirements.txt
psutil
watchdog
requests
pymongo
python-dateutil
python-dotenv
load_dotenv